requirements.txt
loguru
pandas
openpyxl
pytorch
torchvision
pyyaml
tensorboard
tqdm